Consolidate domain architecture (merge models into single models.py)
