+++ backend/app/db/models.py
@@ -1,8 +1,9 @@
 # Import all models to ensure they are registered with SQLAlchemy
-from app.domains.channels.models import Channel  # noqa: F401
-from app.domains.transcription_jobs.models import (  # noqa: F401
+from app.models import (  # noqa: F401
+    Channel,
     JobStatus,
     TranscriptFormat,
     TranscriptionJob,
+    Video,
+    VideoMetadata,
 )
-from app.domains.videos.models import Video  # noqa: F401
