--- backend/app/domains/channels/models.py
@@ -1,38 +1,0 @@
-from __future__ import annotations
-
-from datetime import datetime
-
-from sqlalchemy import DateTime, Integer, String, Text
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-
-from app.db.base import Base
-
-
-class Channel(Base):
-    """YouTube channel model."""
-
-    __tablename__ = "channels"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
-    youtube_id: Mapped[str] = mapped_column(
-        String(255), unique=True, index=True, nullable=False
-    )
-    title: Mapped[str] = mapped_column(String(255), nullable=False)
-    description: Mapped[str | None] = mapped_column(Text, nullable=True)
-    url: Mapped[str] = mapped_column(String(500), nullable=False)
-    thumbnail_url: Mapped[str | None] = mapped_column(String(500), nullable=True)
-    custom_url: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    subscriber_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    video_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    view_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    published_at: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime, default=datetime.utcnow, onupdate=datetime.utcnow
-    )
-
-    # Relationships
-    videos: Mapped[list[Video]] = relationship("Video", back_populates="channel")
-
-    def __repr__(self) -> str:
-        return f"<Channel(id={self.id}, title='{self.title}')>"
