+++ task.md
@@ -13,7 +13,7 @@ How we are building it:
 - [x] Remove pre-commit hooks configuration
 - [x] Clean up unused dependencies from pyproject.toml (yt-dlp, deepgram-sdk, greenlet)
 - [x] Remove Alembic migration system and files (use simple database creation)
-- [ ] Consolidate domain architecture (merge models into single models.py)
+- [x] Consolidate domain architecture (merge models into single models.py)
 - [ ] Remove redundant entry points (keep only run.py, remove main.py and quick_setup.py)
 - [ ] Remove placeholder transcription job routes until actual implementation needed
 - [x] Consolidate duplicate dev dependencies in pyproject.toml
